Finishes implementation of the basic DIMM CSC functionality. (#1)
* Finishes implementation of the basic DIMM CSC functionality.
Adds soar_dimm a controller interface to grab data from the SOAR DIMM database. The controller is still under development, we will access whether or not to finish it in the future. I also need to get access to the database on the summit. The data I have access to at this time is outdated.

* Adds more information regarding the state of SOAR dimm.